Update apple_description_release.lang
First translation draft. I'll share this with the rest of the Spanish-speaking community for a final QA. So you should expect a few more edits.
Additionally, there's no such a thing as an es-ES Firefox for iOS. It's an International Spanish Firefox for iOS so the app description should be under the "es" locale.